requirements.txt
speedtest-cli
requests
rich